Moved equatable conformance into test bundle
