+++ Sources/PiholeNetworking/Models/PHProviderError.swift
@@ -16,20 +16,3 @@ public enum PHProviderError: Error {
 	case urlError(URLError)
 	case other(Error)
 }
-
-extension PHProviderError: Equatable {
-	public static func == (lhs: PHProviderError, rhs: PHProviderError) -> Bool {
-		switch (lhs, rhs) {
-		case (.authenticationRequired, .authenticationRequired):
-			return true
-		case (.invalidHostname, .invalidHostname):
-			return true
-		case (.decodingError(_), .decodingError(_)):
-			return true
-		case (.urlError(let e1),.urlError(let e2)):
-			return e1 == e2
-		default:
-			return false
-		}
-	}
-}
